migrate FirebaseAPI to subfiles
